refactor(whispering): remove transcribe_audio_* compat shims, split audio.rs
Now that the frontend invokes the unified transcribe_audio command,
remove the three per-engine compat commands and parse_moonshine_variant.
The Moonshine variant arrives on the wire as a tagged enum, so the
string scrape of the directory name is gone from Rust entirely.

Move all audio normalization (~400 lines: is_valid_wav_format,
convert_audio_rust, convert_audio_to_pcm16k_mono, convert_audio_with_ffmpeg,
extract_samples_from_wav, prepare_samples_for_transcription) out of
mod.rs into a new audio.rs sibling module. mod.rs is now ~200 lines of
"transcription Tauri command + wire types + engine dispatch" instead
of mostly being audio plumbing.

Also delete src-tauri/src/transcription/audio_converter.rs: it was a
pre-existing orphan file, never declared in any `mod`, never compiled.
