+++ apps/whispering/src-tauri/src/lib.rs
@@ -12,10 +12,7 @@ use recorder::commands::{
 use recorder::recorder::Recorder;
 
 pub mod transcription;
-use transcription::{
-    transcribe_audio, transcribe_audio_moonshine, transcribe_audio_parakeet,
-    transcribe_audio_whisper, ModelManager,
-};
+use transcription::{transcribe_audio, ModelManager};
 
 pub mod windows_path;
 use windows_path::fix_windows_path;
@@ -172,9 +169,6 @@ pub async fn run() {
         stop_recording,
         cancel_recording,
         transcribe_audio,
-        transcribe_audio_whisper,
-        transcribe_audio_parakeet,
-        transcribe_audio_moonshine,
         send_sigint,
         // Command execution (prevents console window flash on Windows)
         execute_command,
